Policy replaces customer/home
